requirements.txt
aiogram>=3.3.0
motor>=3.3.2
redis>=5.0.1
python-dotenv>=1.0.1